# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ApplicationOut": {
        "properties": {
          "coverLetterId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverletterid"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "jobId": {
            "title": "Jobid",
            "type": "integer"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "resumeVersionId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resumeversionid"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "submittedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submittedat"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "id",
          "jobId",
          "status",
          "notes",
          "submittedAt",
          "updatedAt"
        ],
        "title": "ApplicationOut",
        "type": "object"
      },
...
      "CoverLetterOut": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "factCheckPassed": {
            "title": "Factcheckpassed",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "instruction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instruction"
          },
          "jobId": {
            "title": "Jobid",
            "type": "integer"
          },
          "origin": {
            "default": "draft",
            "title": "Origin",
            "type": "string"
          },
          "parentId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parentid"
          },
          "pdfPath": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pdfpath"
          },
          "resumeVersionId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resumeversionid"
          }
        },
        "required": [
          "id",
          "jobId",
          "factCheckPassed",
          "pdfPath",
          "createdAt"
        ],
        "title": "CoverLetterOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LinkTokenRequest": {
        "properties": {
          "purpose": {
            "enum": [
              "sse",
              "download"
            ],
            "title": "Purpose",
            "type": "string"
          }
        },
        "required": [
          "purpose"
        ],
        "title": "LinkTokenRequest",
        "type": "object"
      },
      "LinkTokenResponse": {
        "properties": {
          "expiresInSeconds": {
            "title": "Expiresinseconds",
            "type": "integer"
          },
          "token": {
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token",
          "expiresInSeconds"
        ],
        "title": "LinkTokenResponse",
        "type": "object"
      },
      "LoginRequest": {
        "properties": {
          "password": {
            "maxLength": 1024,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          },
          "username": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
...
      "MeResponse": {
        "properties": {
          "authRequired": {
            "default": false,
            "title": "Authrequired",
            "type": "boolean"
          },
          "role": {
            "anyOf": [
              {
                "enum": [
                  "admin",
                  "user"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          }
        },
        "title": "MeResponse",
        "type": "object"
      },
...
      "RegisterRequest": {
        "properties": {
          "inviteCode": {
            "maxLength": 128,
            "minLength": 8,
            "title": "Invitecode",
            "type": "string"
          },
          "password": {
            "maxLength": 1024,
            "minLength": 12,
            "title": "Password",
            "type": "string"
          },
          "username": {
            "pattern": "^[a-z][a-z0-9._-]{2,31}$",
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password",
          "inviteCode"
        ],
        "title": "RegisterRequest",
        "type": "object"
      },
...
      "ReviseRequest": {
        "properties": {
          "instruction": {
            "title": "Instruction",
            "type": "string"
          },
          "reReview": {
            "default": false,
            "title": "Rereview",
            "type": "boolean"
          }
        },
        "required": [
          "instruction"
        ],
        "title": "ReviseRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Resume Agent API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/auth/link-token": {
      "post": {
        "operationId": "mint_link_token_api_auth_link_token_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkTokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mint Link Token",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/auth/register": {
      "post": {
        "operationId": "register_api_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/cover-letters/{cover_letter_id}/pdf": {
      "get": {
        "operationId": "download_cover_letter_pdf_api_cover_letters__cover_letter_id__pdf_get",
        "parameters": [
          {
            "in": "path",
            "name": "cover_letter_id",
            "required": true,
            "schema": {
              "title": "Cover Letter Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Cover Letter Pdf"
      }
    },
    "/api/cover-letters/{cover_letter_id}/revise": {
      "post": {
        "operationId": "revise_cover_letter_endpoint_api_cover_letters__cover_letter_id__revise_post",
        "parameters": [
          {
            "in": "path",
            "name": "cover_letter_id",
            "required": true,
            "schema": {
              "title": "Cover Letter Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoverLetterOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revise Cover Letter Endpoint"
      }
    },
...
    "/api/jobs/{job_id}/select-cover-letter/{cover_letter_id}": {
      "post": {
        "operationId": "select_cover_letter_endpoint_api_jobs__job_id__select_cover_letter__cover_letter_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "cover_letter_id",
            "required": true,
            "schema": {
              "title": "Cover Letter Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Select Cover Letter Endpoint"
      }
    },
...
  }
}